requirement.txt
python==3.11
tensorflow>=2.13
keras>=2.13
numpy
pillow
opencv-python
scikit-image
scipy
